requirements.txt
pyrogram==2.0.59
py-tgcalls==2.2.5
telethon==1.28.2
python-dotenv
motor
pyromod~=1.5
ffmpeg-python
tgcrypto
youtube-search
youtube-search-python
yt-dlp
dnspython
gitpython
asgiref
aiofiles
aiohttp
requests
pillow
heroku3
psutil
future
speedtest-cli
googletrans